refactor: separate data file and icons folder
