+++ src/components/pages/doc/post/post.jsx
@@ -6,12 +6,12 @@ import Hero from 'components/pages/changelog/hero';
 import Breadcrumbs from 'components/pages/doc/breadcrumbs';
 import EditOnGithub from 'components/pages/doc/edit-on-github';
 import Modal from 'components/pages/doc/modal';
+import MODALS from 'components/pages/doc/modal/data';
 import Content from 'components/shared/content';
 import DocFooter from 'components/shared/doc-footer';
 import NavigationLinks from 'components/shared/navigation-links';
 import TableOfContents from 'components/shared/table-of-contents';
 import { DOCS_BASE_PATH } from 'constants/docs';
-import LINKS from 'constants/links';
 
 import Tag from '../tag';
 
@@ -25,29 +25,6 @@ const Changelog = ({ posts }) => (
 Changelog.propTypes = {
   posts: PropTypes.arrayOf(PropTypes.object).isRequired,
 };
-
-const MODALS = [
-  {
-    type: 'migrate',
-    breadcrumb: 'Migrate to Neon',
-    title: 'Migrating to Neon?',
-    description: 'Our team can help minimize downtime.',
-    link: {
-      title: 'Get migration assistance',
-      url: LINKS.migrationAssistance,
-    },
-  },
-  {
-    type: 'support',
-    breadcrumb: 'Support',
-    title: 'Need help now?',
-    description: 'Please reach out to our Support team!',
-    link: {
-      title: 'Get support',
-      url: LINKS.consoleSupport,
-    },
-  },
-];
 
 const Post = ({
   data: { title, subtitle, enableTableOfContents = false, tag = null, updatedOn = null },
